certchat/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-openai
python-dotenv
faiss-cpu
jinja2
sentence-transformers
pypdf